+++ go/pkg/basecamp/vaults.go
@@ -1096,7 +1096,7 @@ func vaultFromGenerated(gv generated.Vault) Vault {
 
 	if gv.Creator.Id != 0 || gv.Creator.Name != "" {
 		v.Creator = &Person{
-			ID:           gv.Creator.Id,
+			ID:           int64(gv.Creator.Id),
 			Name:         gv.Creator.Name,
 			EmailAddress: gv.Creator.EmailAddress,
 			AvatarURL:    gv.Creator.AvatarUrl,
@@ -1153,7 +1153,7 @@ func documentFromGenerated(gd generated.Document) Document {
 
 	if gd.Creator.Id != 0 || gd.Creator.Name != "" {
 		d.Creator = &Person{
-			ID:           gd.Creator.Id,
+			ID:           int64(gd.Creator.Id),
 			Name:         gd.Creator.Name,
 			EmailAddress: gd.Creator.EmailAddress,
 			AvatarURL:    gd.Creator.AvatarUrl,
@@ -1216,7 +1216,7 @@ func uploadFromGenerated(gu generated.Upload) Upload {
 
 	if gu.Creator.Id != 0 || gu.Creator.Name != "" {
 		u.Creator = &Person{
-			ID:           gu.Creator.Id,
+			ID:           int64(gu.Creator.Id),
 			Name:         gu.Creator.Name,
 			EmailAddress: gu.Creator.EmailAddress,
 			AvatarURL:    gu.Creator.AvatarUrl,
